+++ suite-common/trading/src/types.ts
@@ -88,16 +88,21 @@ type TradingCommonTransaction = {
         accountIndex: Account['index'];
     };
 };
-export type TradingTransactionBuy = TradingCommonTransaction & { tradeType: 'buy'; data: BuyTrade };
+export type TradingTransactionBuy = TradingCommonTransaction & {
+    tradeType: 'buy';
+    data: BuyTrade;
+    receiveAccountKey: Account['key'] | undefined;
+};
 export type TradingTransactionSell = TradingCommonTransaction & {
     tradeType: 'sell';
     data: SellFiatTrade;
+    sendAccountKey: Account['key'] | undefined;
 };
 export type TradingTransactionExchange = TradingCommonTransaction & {
     tradeType: 'exchange';
     data: ExchangeTrade;
-    sendAccountKey?: Account['key'];
     receiveAccountKey?: Account['key'];
+    sendAccountKey: Account['key'] | undefined;
 };
 export type TradingTransaction =
     | TradingTransactionBuy